requirements.txt
discord.py==0.16.12
beem==0.20.18
requests==2.21.0
